Update bgraknob.pas minor fix for default property values
Ran into an issue where some property values were not being saved in the .lfm, notably when the min/max values were 0. This made the default values come from the Create method which is only a fall back if not set by the design time props. Set property to 'nodefault' seems to force it to always write. Oddly only happened with the floating point type when the value was 0.